move to qt/widgets/replay_widgets
