+++ selfdrive/ui/qt/onroad.h
@@ -1,37 +1,13 @@
 #pragma once
 
-#include <QDateTime>
-#include <QLabel>
 #include <QStackedLayout>
-#include <QPushButton>
-#include <QTimer>
 #include <QWidget>
-#include <QSlider>
 
 #include "common/util.h"
 #include "selfdrive/ui/qt/widgets/cameraview.h"
+#include "selfdrive/ui/qt/widgets/replay_widgets.h"
 #include "selfdrive/ui/ui.h"
-#include "tools/replay/replay.h"
 
-class ReplayControls : public QWidget {
-  Q_OBJECT
-
-public:
-  ReplayControls(QWidget *parent);
-  void start(const QString &route, const QString &data_dir);
-  void adjustPosition();
-
-protected:
-  inline QString formatTime(int seconds) {
-    return QDateTime::fromTime_t(seconds).toString(seconds > 60 * 60 ? "hh:mm:ss" : "mm:ss");
-  }
-  QSlider *slider;
-  QLabel *end_time_label;
-  QPushButton *play_btn;
-  QPushButton *stop_btn;
-  QTimer *timer;
-  std::unique_ptr<Replay> replay;
-};
 
 // ***** onroad widgets *****
 class OnroadAlerts : public QWidget {
